Fix void function error with advice version for sx.
Fixes #5.
The pre-#336 version of sx.el does not have sx-question-mode-after-print-hook.
It must use a function advice for sx-question-mode--print-question.
But that cannot be applied buffer-locally.

Therefore, if one switches texfrag on in one buffer texfrag-sx-after-print
is also called in other buffers where texfrag-mode is not properly setup.
Especially texfrag-preview-region-function is nil.